+++ src/entities/chat-item/index.ts
@@ -1,0 +1,1 @@
+export { default as ChatItem } from './ui/chat-item.component';
